% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_utils.R
\name{df_to_spectra}
\alias{df_to_spectra}
\title{Convert data.frame back to Spectra object}
\usage{
df_to_spectra(df)
}
\arguments{
\item{df}{A data.frame with metadata columns plus \code{mz} and \code{intensity}
as list columns}
}
\value{
A Spectra object
}
\description{
Reconstructs a Spectra object from a portable data.frame format created by
\code{spectra_to_df}. The data.frame must have \code{mz} and \code{intensity} columns
(as lists) along with metadata columns.
}
\examples{
\dontrun{
  spectra_obj <- Spectra::Spectra(mgf_file, source = "MgfSqlite")
  df <- spectra_to_df(spectra_obj)
  spectra_obj_restored <- df_to_spectra(df)
  identical(length(spectra_obj), length(spectra_obj_restored))
}

}
