% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enveda180_neg}
\alias{enveda180_neg}
\title{Enveda-180 Spectral Library - Negative Polarity}
\format{
Same as \code{enveda180_pos}
}
\source{
\doi{10.5281/zenodo.20436851}
}
\usage{
enveda180_neg
}
\description{
A data.frame containing metadata and spectra from Enveda-180 (negative polarity).
}
\examples{
data(enveda180_neg)
head(enveda180_neg)
}
\keyword{datasets}
