% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enveda180_pos}
\alias{enveda180_pos}
\title{Enveda-180 Spectral Library - Positive Polarity}
\format{
A data.frame with one row per spectrum
}
\source{
\doi{10.5281/zenodo.20436851}
}
\usage{
enveda180_pos
}
\description{
A data.frame containing metadata and spectra from Enveda-180 (positive polarity),
a curated MS/MS spectral library. Peaks are stored as list columns.
}
\examples{
data(enveda180_pos)
head(enveda180_pos)
}
\keyword{datasets}
