% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enveda180_prepared}
\alias{enveda180_prepared}
\title{Enveda-180 Prepared Table}
\format{
A data.frame prepared by \code{tima::prepare_libraries_spectra()}.
}
\source{
\doi{10.5281/zenodo.20436851}
}
\usage{
enveda180_prepared
}
\description{
A data.frame containing the prepared tabular Enveda-180 metadata and
spectrum annotations used to derive the shipped Enveda-180 polarity-split
datasets.
}
\examples{
data(enveda180_prepared)
head(enveda180_prepared)
}
\keyword{datasets}
