% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gnps_neg}
\alias{gnps_neg}
\title{GNPS Spectral Library - Negative Polarity}
\format{
Same as \code{gnps_pos}
}
\source{
\doi{10.5281/zenodo.11566051}
}
\usage{
gnps_neg
}
\description{
A data.frame containing metadata and spectra from the Global Natural Products
Social Molecular Networking (GNPS) database (negative polarity). Peaks (m/z
and intensity values) are stored as list columns for compatibility with Spectra
objects.
}
\examples{
data(gnps_neg)
head(gnps_neg)
nrow(gnps_neg)
}
\keyword{datasets}
