% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gnps_prepared}
\alias{gnps_prepared}
\title{GNPS Prepared Table}
\format{
A data.frame prepared by \code{tima::prepare_libraries_spectra()}.
}
\source{
\doi{10.5281/zenodo.11566051}
}
\usage{
gnps_prepared
}
\description{
A data.frame containing the prepared tabular GNPS metadata and spectrum
annotations used to derive the shipped GNPS polarity-split datasets.
}
\examples{
data(gnps_prepared)
head(gnps_prepared)
}
\keyword{datasets}
