% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isdbnormansusdat_neg}
\alias{isdbnormansusdat_neg}
\title{ISDB NormanSusDat Spectral Library - Negative Polarity}
\format{
Same as \code{isdbnormansusdat_pos}
}
\source{
\doi{10.5281/zenodo.14854025}
}
\usage{
isdbnormansusdat_neg
}
\description{
A data.frame containing metadata and spectra from ISDB NormanSusDat
(negative polarity).
}
\examples{
data(isdbnormansusdat_neg)
head(isdbnormansusdat_neg)
}
\keyword{datasets}
