% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isdbnormansusdat_pos}
\alias{isdbnormansusdat_pos}
\title{ISDB NormanSusDat Spectral Library - Positive Polarity}
\format{
A data.frame with one row per spectrum
}
\source{
\doi{10.5281/zenodo.14854025}
}
\usage{
isdbnormansusdat_pos
}
\description{
A data.frame containing metadata and spectra from ISDB NormanSusDat
(positive polarity), a comprehensive MS/MS database. Peaks are stored as
list columns.
}
\examples{
data(isdbnormansusdat_pos)
head(isdbnormansusdat_pos)
}
\keyword{datasets}
