% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isdbnormansusdat_prepared}
\alias{isdbnormansusdat_prepared}
\title{ISDB NormanSusDat Prepared Table}
\format{
A data.frame prepared by \code{tima::prepare_libraries_spectra()}.
}
\source{
\doi{10.5281/zenodo.14854025}
}
\usage{
isdbnormansusdat_prepared
}
\description{
A data.frame containing the prepared tabular ISDB NormanSusDat metadata and
spectrum annotations used to derive the shipped polarity-split datasets.
}
\examples{
data(isdbnormansusdat_prepared)
head(isdbnormansusdat_prepared)
}
\keyword{datasets}
