% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isdbwikidata_neg}
\alias{isdbwikidata_neg}
\title{ISDB Wikidata Spectral Library - Negative Polarity}
\format{
Same as \code{isdbwikidata_pos}
}
\source{
\doi{10.5281/zenodo.5607185}
}
\usage{
isdbwikidata_neg
}
\description{
A data.frame containing metadata and spectra from ISDB Wikidata
(negative polarity).
}
\examples{
data(isdbwikidata_neg)
head(isdbwikidata_neg)
}
\keyword{datasets}
