% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isdbwikidata_pos}
\alias{isdbwikidata_pos}
\title{ISDB Wikidata Spectral Library - Positive Polarity}
\format{
A data.frame with one row per spectrum
}
\source{
\doi{10.5281/zenodo.5607185}
}
\usage{
isdbwikidata_pos
}
\description{
A data.frame containing metadata and spectra from ISDB Wikidata
(positive polarity), a community-curated MS/MS spectral database.
Peaks are stored as list columns.
}
\examples{
data(isdbwikidata_pos)
head(isdbwikidata_pos)
}
\keyword{datasets}
