% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{massbank_neg}
\alias{massbank_neg}
\title{MassBank Spectral Library - Negative Polarity}
\format{
Same as \code{massbank_pos}
}
\source{
MassBank database. See \url{https://www.massbank.jp/}
}
\usage{
massbank_neg
}
\description{
A data.frame containing metadata and spectra from MassBank (negative polarity).
}
\examples{
data(massbank_neg)
head(massbank_neg)
}
\keyword{datasets}
