% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{massbank_prepared}
\alias{massbank_prepared}
\title{MassBank Prepared Table}
\format{
A data.frame prepared by \code{tima::prepare_libraries_spectra()}.
}
\source{
MassBank database. See \url{https://www.massbank.jp/}
}
\usage{
massbank_prepared
}
\description{
A data.frame containing the prepared tabular MassBank metadata and spectrum
annotations used to derive the shipped MassBank polarity-split datasets.
}
\examples{
data(massbank_prepared)
head(massbank_prepared)
}
\keyword{datasets}
