% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merlin_neg}
\alias{merlin_neg}
\title{Merlin Spectral Library - Negative Polarity}
\format{
Same as \code{merlin_pos}
}
\source{
Merlin library (Zenodo)
}
\usage{
merlin_neg
}
\description{
A data.frame containing metadata and spectra from Merlin (negative polarity).
}
\examples{
data(merlin_neg)
head(merlin_neg)
}
\keyword{datasets}
