% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merlin_pos}
\alias{merlin_pos}
\title{Merlin Spectral Library - Positive Polarity}
\format{
A data.frame with one row per spectrum and columns similar to GNPS/MassBank
}
\source{
Merlin library (Zenodo)
}
\usage{
merlin_pos
}
\description{
A data.frame containing metadata and spectra from Merlin (positive polarity),
a high-quality MS/MS spectral library. Peaks are stored as list columns.
}
\examples{
data(merlin_pos)
head(merlin_pos)
}
\keyword{datasets}
