% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multims2_neg}
\alias{multims2_neg}
\title{MultiMS2 Spectral Library - Negative Polarity}
\format{
Same as \code{multims2_pos}
}
\source{
\doi{10.5281/zenodo.17417089}
}
\usage{
multims2_neg
}
\description{
A data.frame containing metadata and spectra from MultiMS2 (negative polarity).
}
\examples{
data(multims2_neg)
head(multims2_neg)
}
\keyword{datasets}
