% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multims2_prepared}
\alias{multims2_prepared}
\title{MultiMS2 Prepared Table}
\format{
A data.frame prepared by \code{tima::prepare_libraries_spectra()}.
}
\source{
\doi{10.5281/zenodo.17417089}
}
\usage{
multims2_prepared
}
\description{
A data.frame containing the prepared tabular MultiMS2 metadata and spectrum
annotations used to derive the shipped MultiMS2 polarity-split datasets.
}
\examples{
data(multims2_prepared)
head(multims2_prepared)
}
\keyword{datasets}
